make some files and moved some
