+++ src/Viserio/Component/Routing/Route/Group.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Component\Routing\Route;
+
+class Group
+{
+
+}
